Move SystemHealth to its own crate.
